complete ver 1.0

+++ main.py
@@ -1,21 +1,33 @@
 import os
 import shutil
 import argparse
+import codecs
 
 
 ALT_SEPARATOR = '$'
 
 
-def transform(src, dst, log):
+def transform(src, dst, log, extensions):
+    existed_files = list()
+    if log is not None and os.path.exists(log):
+        for line in codecs.open(log, 'r', 'utf8'):
+            line = line.strip()
+            existed_files.append(line)
+    writer = codecs.open(log, 'a', 'utf8')
     for root, dirs, files in os.walk(src):
         for f in files:
-            src_path = os.path.abspath(os.path.join(root, f))
-            prefix = os.path.abspath(src)
-            file_name = src_path[len(prefix)+1:]
-            dst_path = dst + os.sep + file_name.replace(os.sep, ALT_SEPARATOR)
-            print 'Source: ', src_path
-            print 'Destination: ', dst_path
-            shutil.copyfile(src_path, dst_path)
+            filename, extension = os.path.splitext(f)
+            if extensions == 'all' or extension in extensions:
+                src_path = os.path.abspath(os.path.join(root, f))
+                prefix = os.path.abspath(src)
+                file_name = src_path[len(prefix)+1:]
+                dst_path = dst + os.sep + file_name.replace(os.sep, ALT_SEPARATOR)
+                if file_name not in existed_files:
+                    print 'Source: ', src_path
+                    print 'Destination: ', dst_path
+                    shutil.copyfile(src_path, dst_path)
+                    writer.write(file_name + '\n')
+    writer.close()
 
 
 def mkdir_safe(path):
@@ -30,7 +42,7 @@ def mkdir_safe(path):
             os.mkdir(d)
 
 
-def convert(src, dst, log):
+def convert(src, dst):
     for root, dirs, files in os.walk(src):
         for f in files:
             src_path = os.path.join(root, f)
@@ -40,7 +52,8 @@ def convert(src, dst, log):
             if not os.path.exists(dir_path):
                 print 'Path: ', dir_path
                 mkdir_safe(dir_path)
-            shutil.copyfile(src_path, dst_path)
+            if not os.path.exists(dst_path):
+                shutil.copyfile(src_path, dst_path)
 
 
 def main():
@@ -48,15 +61,17 @@ def main():
     parser.add_argument('-s', '--src', required=True, help='Source folder')
     parser.add_argument('-d', '--dst', required=True, help='Destination folder')
     parser.add_argument('-m', '--mode', default='transform', help='Mode: transform/convert')
-    parser.add_argument('-e', '--ext', default='all', help='File extension')
+    parser.add_argument('-e', '--ext', default='all', nargs='+', help='File extension')
     parser.add_argument('-l', '--list', default=None, help='list')
 
     args = parser.parse_args()
-    src = '/home/khoaipx/chatbot_log'
-    dst = '/home/khoaipx/proj'
-    mode = ''
-    # transform('/home/khoaipx/chatbot_log', '/home/khoaipx/proj/test', None)
-    # convert('/home/khoaipx/proj/test', '/home/khoaipx/proj/test2', None)
+
+    if args.mode == 'transform':
+        transform(args.src, args.dst, args.list, args.ext)
+    elif args.mode == 'convert':
+        convert(args.src, args.dst)
+    else:
+        print 'Program only supports 2 modes: transform/convert'
 
 
 if __name__ == '__main__':
